KRMVP-863 Error when removing device with a large number of events
