+++ konker.registry.data/src/main/java/com/konkerlabs/platform/registry/integration/endpoints/DeviceConfigRabbitEndpoint.java
@@ -15,7 +15,7 @@ import com.konkerlabs.platform.registry.business.model.Device;
 import com.konkerlabs.platform.registry.business.services.api.DeviceConfigSetupService;
 import com.konkerlabs.platform.registry.business.services.api.DeviceRegisterService;
 import com.konkerlabs.platform.registry.business.services.api.ServiceResponse;
-import com.konkerlabs.platform.registry.data.config.RabbitMQConfig;
+import com.konkerlabs.platform.registry.data.config.RabbitMQDataConfig;
 import com.konkerlabs.platform.registry.integration.gateways.RabbitGateway;
 
 @Service
@@ -56,7 +56,7 @@ public class DeviceConfigRabbitEndpoint {
             return;
         }
 
-        String apiKey = (String) properties.getHeaders().get(RabbitMQConfig.MSG_HEADER_APIKEY);
+        String apiKey = (String) properties.getHeaders().get(RabbitMQDataConfig.MSG_HEADER_APIKEY);
         if (!StringUtils.hasText(apiKey)) {
             LOGGER.error("Apikey not found.");
             return;
