+++ konker.registry.data/src/main/java/com/konkerlabs/platform/registry/integration/gateways/RabbitGateway.java
@@ -11,7 +11,7 @@ import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.konkerlabs.platform.registry.data.config.RabbitMQConfig;
+import com.konkerlabs.platform.registry.data.config.RabbitMQDataConfig;
 
 @Service
 public class RabbitGateway {
@@ -25,8 +25,8 @@ public class RabbitGateway {
 
         try {
             MessageProperties properties = new MessageProperties();
-            properties.setHeader(RabbitMQConfig.MSG_HEADER_APIKEY, apiKey);
-            properties.setHeader(RabbitMQConfig.MSG_HEADER_CHANNEL, channel);
+            properties.setHeader(RabbitMQDataConfig.MSG_HEADER_APIKEY, apiKey);
+            properties.setHeader(RabbitMQDataConfig.MSG_HEADER_CHANNEL, channel);
 
             Message message = new Message(payload, properties);
 
@@ -40,7 +40,7 @@ public class RabbitGateway {
 
         try {
             MessageProperties properties = new MessageProperties();
-            properties.setHeader(RabbitMQConfig.MSG_HEADER_APIKEY, apiKey);
+            properties.setHeader(RabbitMQDataConfig.MSG_HEADER_APIKEY, apiKey);
 
             Message message = new Message(config.getBytes("UTF-8"), properties);
