+++ konker.registry.services.core/pom.xml
@@ -51,6 +51,12 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
+		<!-- RabbitMQ -->
+        <dependency>
+            <groupId>org.springframework.integration</groupId>
+            <artifactId>spring-integration-amqp</artifactId>
+            <version>4.3.9.RELEASE</version>
+        </dependency>
         <!-- Apache -->
         <dependency>
             <groupId>commons-codec</groupId>
