+++ konker.registry.web/src/main/java/com/konkerlabs/platform/registry/RegistryAppInitializer.java
@@ -18,6 +18,7 @@ import com.konkerlabs.platform.registry.config.CdnConfig;
 import com.konkerlabs.platform.registry.config.EmailConfig;
 import com.konkerlabs.platform.registry.config.MessageSourceConfig;
 import com.konkerlabs.platform.registry.config.PubServerConfig;
+import com.konkerlabs.platform.registry.config.RabbitMQConfig;
 import com.konkerlabs.platform.registry.config.WebMvcConfig;
 
 public class RegistryAppInitializer
@@ -43,6 +44,7 @@ public class RegistryAppInitializer
             com.konkerlabs.platform.registry.config.AmazonConfig.class,
             com.konkerlabs.platform.registry.config.EnvironmentConfig.class,
             com.konkerlabs.platform.registry.config.EventStorageConfig.class,
+            RabbitMQConfig.class,
             MessageSourceConfig.class
             
         };
